Implement HW Lesson 1. v. 1

+++ src/MainArray.java
@@ -33,11 +33,19 @@ public class MainArray {
                 case "save":
                     r = new Resume();
                     r.uuid = uuid;
-                    ARRAY_STORAGE.save(r);
-                    printAll();
+                    if(ARRAY_STORAGE.get(uuid) != null) {
+                        System.out.println("Резюме с id " + uuid + " уже существует");
+                    } else {
+                        ARRAY_STORAGE.save(r);
+                        printAll();
+                    }
                     break;
                 case "delete":
-                    ARRAY_STORAGE.delete(uuid);
+                    if(ARRAY_STORAGE.get(uuid) == null) {
+                        System.out.println("Резюме с id " + uuid + " не найдено");
+                    } else {
+                        ARRAY_STORAGE.delete(uuid);
+                    }
                     printAll();
                     break;
                 case "get":
